scripts/requirements.txt
requests
python-lorem
